split python and arrow
