+++ src/export/mod.rs
@@ -1,5 +1,6 @@
 //! Module to export mdf files to other file formats.
 pub mod numpy;
+pub mod python_arrow_helpers;
 //pub mod parquet;
 pub mod polars;
 // pub mod tensor;
